Extract environment variable
