+++ src/main/java/com/yam/app/account/infrastructure/MailManager.java
@@ -1,8 +1,7 @@
-package com.yam.app.account.application;
+package com.yam.app.account.infrastructure;
 
 import com.yam.app.account.domain.RegisterAccountEvent;
-import com.yam.app.account.infrastructure.MailDispatcher;
-import com.yam.app.account.infrastructure.MailMessage;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 import org.thymeleaf.TemplateEngine;
@@ -11,14 +10,15 @@ import org.thymeleaf.context.Context;
 @Component
 final class MailManager {
 
-    private static final String HOST = "http://localhost:8080";
-
     private final MailDispatcher mailDispatcher;
     private final TemplateEngine templateEngine;
+    private final String host;
 
-    public MailManager(MailDispatcher mailDispatcher, TemplateEngine templateEngine) {
+    public MailManager(@Value("${app.mail.host}") String host,
+        MailDispatcher mailDispatcher, TemplateEngine templateEngine) {
         this.mailDispatcher = mailDispatcher;
         this.templateEngine = templateEngine;
+        this.host = host;
     }
 
     @EventListener
@@ -31,7 +31,7 @@ final class MailManager {
         context.setVariable("nickname", newAccount.getNickname());
         context.setVariable("linkName", "이메일 인증하기");
         context.setVariable("message", "YouAndMe 서비스를 사용하려면 링크를 클릭하세요.");
-        context.setVariable("host", HOST);
+        context.setVariable("host", host);
         String message = templateEngine.process("mail/check-token", context);
 
         var mailMessage = MailMessage.builder()
