+++ src/main/resources/application.yml
@@ -10,3 +10,7 @@ spring:
         timeout: 5000
         starttls:
           enable: true
+
+app:
+  mail:
+    host: http://localhost:8080
